# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Kite API",
		"version": "2026.10.09.0",
		"description": "Generated from src/app/api/v1/**/route.ts by scripts/docs-generate.mjs. See docs/api.md."
	},
	"servers": [{ "url": "/" }],
	"paths": {
...
		"/api/v1/domains/{id}/dns/setup": {
			"post": {
				"operationId": "postDomainsIdDnsSetup",
				"security": [{ "bearerAuth": ["domains"] }],
				"x-requires-admin": true,
				"parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
				"requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
				"responses": {
					"200": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } },
					"400": { "$ref": "#/components/responses/Error" },
					"401": { "$ref": "#/components/responses/Error" },
					"403": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" },
					"409": { "$ref": "#/components/responses/Error" },
					"500": { "$ref": "#/components/responses/Error" }
				}
			}
		},
		"/api/v1/domains/{id}": {
			"delete": {
				"operationId": "deleteDomainsId",
				"security": [{ "bearerAuth": ["domains"] }],
				"x-requires-admin": true,
				"parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
				"responses": {
					"200": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } },
					"400": { "$ref": "#/components/responses/Error" },
					"401": { "$ref": "#/components/responses/Error" },
					"403": { "$ref": "#/components/responses/Error" }
				}
			},
			"get": {
				"operationId": "getDomainsId",
				"security": [{ "bearerAuth": ["domains"] }],
				"x-requires-admin": true,
				"parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
				"responses": {
					"200": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } },
					"401": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" }
				}
			}
		},
...
		"/api/v1/messages": {
			"get": {
				"operationId": "getMessages",
				"security": [{ "bearerAuth": ["read"] }],
				"responses": {
					"200": { "description": "Success", "content": { "application/json": { "schema": { "type": "object" } } } },
					"400": { "$ref": "#/components/responses/Error" },
					"401": { "$ref": "#/components/responses/Error" },
					"404": { "$ref": "#/components/responses/Error" }
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"description": "API key from Settings → API keys (personal scopes: send, read, jmap, calendar:read, calendar:write) or Admin → API keys (admin scopes: domains, accounts, mailboxes; owner must keep the admin role). Scope lists: src/lib/api/scopes.ts."
			}
		},
		"responses": {
			"Error": {
				"description": "Error. Validation failures return the flattened Zod error in `error`.",
				"content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": { "oneOf": [{ "type": "string" }, { "$ref": "#/components/schemas/ValidationError" }] },
					"code": { "type": "string" },
					"records": { "type": "array", "items": {} }
				}
			},
			"ValidationError": {
				"type": "object",
				"properties": {
					"formErrors": { "type": "array", "items": { "type": "string" } },
					"fieldErrors": {
						"type": "object",
						"additionalProperties": { "type": "array", "items": { "type": "string" } }
					}
				}
			}
		}
	}
}